update: change schema module to be schema package, and then organize each schema class into separate modules
