+++ app/schema/__init__.py
@@ -1,0 +1,3 @@
+from .user import *
+from .category import *
+from .todo import *
