+++ app/schema/base.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+from typing import List, TypeVar
+
+T = TypeVar('T')
+
+
+class ResponseModel(BaseModel):
+    message: str
+    status: str
+    data: List[T]
