+++ app/schema/category.py
@@ -1,0 +1,12 @@
+from .base import BaseModel, ResponseModel
+from .user import UserSchema
+
+
+class CategorySchema(BaseModel):
+    id: int
+    name: str
+    user: UserSchema
+
+
+class ListCategoryResponse(ResponseModel):
+    pass
