+++ app/schema/user.py
@@ -1,0 +1,10 @@
+from .base import BaseModel, ResponseModel
+
+
+class UserSchema(BaseModel):
+    id: int
+    name: str
+
+
+class ListUserSchema(ResponseModel):
+    pass
